requirements.txt
absl-py==0.12.0
argon2-cffi==21.1.0
asgiref==3.4.1
astunparse==1.6.3
attrs==21.2.0
Automat==20.2.0
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==4.1.0
blis==0.7.4
cachetools==4.2.2
catalogue==2.0.4
certifi==2020.12.5
cffi==1.15.0
chardet==4.0.0
ci-info==0.2.0
click==7.1.2
configobj==5.0.6
configparser==5.0.2
constantly==15.1.0
cryptography==35.0.0
cssselect==1.1.0
cupy-cuda113==10.0.0
cycler==0.10.0
cymem==2.0.5
Cython==0.29.23
debugpy==1.5.1
decorator==4.4.2
defusedxml==0.7.1
entrypoints==0.3
et-xmlfile==1.1.0
etelemetry==0.2.2
fastrlock==0.8
filelock==3.0.12
fitz==0.0.1.dev2
flatbuffers==1.12
future==0.18.2
gast==0.4.0
google-auth==1.30.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.34.1
h2==3.2.0
h5py==3.1.0
hpack==3.0.0
htmlmin==0.1.12
httplib2==0.19.1
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
ImageHash==4.2.1
incremental==21.3.0
ipykernel==6.5.1
ipython==7.29.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
isodate==0.6.0
itemadapter==0.4.0
itemloaders==1.0.4
jedi==0.18.1
Jinja2==3.0.0
jmespath==0.10.0
joblib==1.0.1
jsonschema==4.2.1
jupyter==1.0.0
jupyter-client==7.1.0
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keras==2.7.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
libclang==12.0.0
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.3
missingno==0.5.0
mistune==0.8.4
multimethod==1.6
murmurhash==1.0.5
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.1
netron==5.0.7
networkx==2.5.1
nibabel==3.2.1
nipype==1.6.1
notebook==6.4.6
numpy==1.19.5
oauthlib==3.1.0
opencv-python==4.5.2.52
openpyxl==3.0.7
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.4
pandas-profiling==3.1.0
pandocfilters==1.5.0
parsel==1.6.0
parso==0.8.2
pathlib==1.0.1
pathy==0.5.2
patsy==0.5.2
pdfminer==20191125
pdfminer.six==20200517
pdfplumber==0.5.28
pexpect==4.8.0
phik==0.12.0
pickleshare==0.7.5
Pillow==8.2.0
pip-autoremove==0.10.0
preshed==3.0.5
priority==1.3.0
prometheus-client==0.12.0
prompt-toolkit==3.0.22
Protego==0.1.16
protobuf==3.16.0
prov==2.0.0
ptyprocess==0.7.0
PuLP==2.6.0
pyaml==21.10.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.10.1
pydantic==1.8.2
PyDispatcher==2.0.5
pydot==1.4.2
pyenchant==3.2.0
Pygments==2.10.0
PyMuPDF==1.18.14
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.2.0
pyxnat==1.4
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.2.1
QtPy==1.11.2
queuelib==1.6.2
rdflib==5.0.0
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==1.0.2
scikit-optimize==0.9.0
scipy==1.6.3
Scrapy==2.5.1
seaborn==0.11.1
Send2Trash==1.8.0
service-identity==21.1.0
simplejson==3.17.2
six==1.15.0
smart-open==3.0.0
sortedcontainers==2.4.0
soupsieve==2.2.1
spacy==3.0.6
spacy-legacy==3.0.5
spacy-pkuseg==0.0.28
sqlparse==0.4.1
srsly==2.4.1
statsmodels==0.13.0
tangled-up-in-unicode==0.1.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.7.0
tensorflow-gpu==2.7.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.22.0
tensorflow-text==2.5.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
thinc==8.0.3
threadpoolctl==2.1.0
tornado==6.1
traitlets==5.1.1
traits==6.2.0
Twisted==21.7.0
typer==0.3.2
typing-extensions==3.7.4.3
urllib3==1.25.8
visions==0.7.4
w3lib==1.22.0
Wand==0.6.6
wasabi==0.8.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.0
widgetsnbextension==3.5.2
wrapt==1.12.1
xgboost==1.5.1
zope.interface==5.4.0

tensorly>=0.7.0